Cleaning and Wrapping Up
Think it's time to clean this up, I've decided to work on something else and call this ready for more user-friendly productive feedback.

I would consider this ready for use, for now. It encompasses all that I wanted for simple SV usage and enough to really help the general users

It has been an enriching experience doing this and I think I can do something better than this, so stay tuned.